It fucking works NOW

+++ src/private/common_utils.h
@@ -10,8 +10,8 @@ typedef enum {
 	UNI_WRITE_TARGET_STDERR = 2
 } UniWriteTarget;
 
-int UniWrite(UniWriteTarget target, const void* buf, size_t n);
-int UniWriteLen(UniWriteTarget target, const void* buf);
+size_t UniWrite(UniWriteTarget target, const void* buf, size_t n);
+size_t UniWriteLen(UniWriteTarget target, const void* buf);
 size_t WriteToBackbuff(const SBCell* content, size_t cellCount);
 
 void EnableRawMode(void);
